fix(search,queries/spans,join): reconcile against Apache Lucene 10.5.0
search carried 2294 compile errors; it now carries none. The work was
reconciliation against the Java source, not repair in place.

Core contracts. Query declared Clone() Query - org.apache.lucene.search.
Query has no clone(), queries are immutable - and CreateWeight took
needsScores bool where Java takes ScoreMode. Both corrected; 45 fabricated
Clone methods removed and one kept (bkdPointTreeWalker.Clone ports
PointValues.PointTree.clone). Scorer, Scorable, ScorerSupplier, SimScorer,
LeafCollector, Collector, Weight, Similarity and SpanQuery brought onto
their Java member sets, with Base* types carrying exactly Java's concrete
defaults. ConstantScoreWeight restored to Java's abstract base class - 14
of its 15 callers already expected that shape.

Duplicate declarations. 62 redeclarations resolved by keeping the
definition in the file named after the Lucene class, after diffing both
copies against Java. Where a name denoted two genuinely distinct Java
artefacts (sigmoid, ToAutomaton, RewriteMethod, dummyImpacts,
scoreCachingLeafCollector), both behaviours were kept and the identifier
disambiguated - in the ScoreCachingWrapping case by restoring the two
distinct Java names.

Misplaced packages. Lucene 10.5.0 has no org.apache.lucene.search.spans:
all 28 spans classes live in queries/spans. The 22-file search/span_*
family duplicated the already-faithful queries/spans, so the 9 missing
classes were ported there - SpanOrQuery, SpanNotQuery, SpanFirstQuery,
SpanPositionCheckQuery, SpanPositionRangeQuery, SpanContainingQuery,
SpanWithinQuery, FieldMaskingSpanQuery and the nested SpanContainWeight -
and the duplicates removed, with 44 consumers repointed. Likewise the four
search/global_ordinals*, terms_including_score_query files, which belong
to join; join/global_ordinals_collector.go was a map[int]int stub and is
replaced by the faithful port that had been sitting in search.

Subsystems ported. FuzzyQuery and FuzzyTermsEnum (the file had been gutted
456 -> 35 lines), MultiTermQuery, TopTermsRewrite, BoostAttribute,
SingleTermsEnum, SloppyPhraseMatcher (all 24 Java members), PhraseWeight,
PhraseScorer, PhrasePositions, PhraseQueue, TermWeight, TermScorer,
CombinedFieldQuery, FieldExistsQuery, IndexSearcher's generic search path
and SortedSetSelector.

Measured: TopTermsBlendedFreqScoringRewrite now orders by boost rather
than enumeration order - top 4 by boost [lucene lucane lucenee luceno]
versus first 4 by enumeration [lucane lucen lucene lucenee].

Behaviour defects fixed in passing: BooleanQuery.contains compared by
identity where Java's Set compares by equals, so MUST_NOT-contains-
MatchAllDocsQuery never fired; ClassicSimScorer passed (docCount, docFreq)
to an idf(docFreq, docCount); the QueryVisitor was inverted in
BayesianScoreQuery and SpanNotQuery; FloatRangeSlowRangeQuery seeded
Arrays.hashCode at 0 instead of 1 and compared raw floats rather than
floatToIntBits.

Around 20 live stubs removed and replaced with the Java body, among them
IndexSearcher.searchQuery and searchWeight, AcceptDocs.NewBitSetIterator,
TermWeight.getCollectionStats, TermInSetQuery.asByteRunAutomaton,
LateInteractionFloatValuesSource.getValues and
SloppyPhraseMatcher.advanceRpts.

+++ search/ib_similarity.go
@@ -74,7 +74,7 @@ func NewIBSimilarityWithDiscount(discountOverlaps bool, distribution IBDistribut
 		normExpl := ib.normalization.Explain(stats, freq, docLen)
 		lambdaExpl := ib.lambda.Explain(stats)
 		subs = append(subs, normExpl, lambdaExpl)
-		subs = append(subs, ib.distribution.Explain(stats, normExpl.GetValue(), lambdaExpl.GetValue()))
+		subs = append(subs, ib.distribution.Explain(stats, float64(normExpl.GetValue()), float64(lambdaExpl.GetValue())))
 		return subs
 	}
